# Conflicts:
#	features/operational/contact_us/build.gradle
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/di/InboxModule.kt
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/domain/usecase/UploadImageUseCase.kt
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/view/activity/InboxDetailActivity.kt
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/view/adapter/AttachmentAdapter.kt
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/view/adapter/InboxDetailAdapter.kt
#	features/operational/contact_us/src/main/java/com/tokopedia/contactus/inboxticket2/view/presenter/InboxDetailPresenterImpl.kt

+++ features/fintech/saldo_details/src/main/java/com/tokopedia/saldodetails/response/model/DepositHistoryList.kt
@@ -10,7 +10,7 @@ class DepositHistoryList : ParcelableViewModel<SaldoDetailTransactionFactory> {
     var note: String? = null
 
     @SerializedName("amount")
-    var amount: Float = 0.toFloat()
+    var amount: Double = 0.0
 
     @SerializedName("image")
     var imageURL: String? = null
