+++ features/transaction/cart/src/main/java/com/tokopedia/cart/data/model/response/shopgroupsimplified/PriceChanges.kt
@@ -7,13 +7,13 @@ import com.google.gson.annotations.SerializedName
 data class PriceChanges(
 
 	@field:SerializedName("amount_difference")
-	val amountDifference: Int = 0,
+	val amountDifference: Long = 0,
 
 	@field:SerializedName("changes_state")
 	val changesState: Int = 0,
 
 	@field:SerializedName("original_amount")
-	val originalAmount: Int = 0,
+	val originalAmount: Long = 0,
 
 	@field:SerializedName("description")
 	val description: String = ""
